backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
httpx>=0.28
websockets>=14.0
python-jose[cryptography]>=3.3
pydantic-settings>=2.7
python-multipart>=0.0.18
